+++ server/core_tournament.go
@@ -69,11 +69,14 @@ func TournamentDelete(ctx context.Context, logger *zap.Logger, cache Leaderboard
 		// If it does not exist treat it as success.
 		return nil
 	}
-	// Allow deletion of non-tournament leaderboards here.
 
 	var expiryUnix int64
 	if leaderboard.ResetSchedule != nil {
 		expiryUnix = leaderboard.ResetSchedule.Next(time.Now().UTC()).UTC().Unix()
+	}
+
+	if leaderboard.EndTime > 0 && expiryUnix > leaderboard.EndTime {
+		expiryUnix = leaderboard.EndTime
 	}
 
 	if err := cache.Delete(ctx, leaderboardId); err != nil {
@@ -104,6 +107,9 @@ func TournamentAddAttempt(ctx context.Context, logger *zap.Logger, db *sql.DB, c
 	expiryTime := int64(0)
 	if leaderboard.ResetSchedule != nil {
 		expiryTime = leaderboard.ResetSchedule.Next(time.Now().UTC()).UTC().Unix()
+		if leaderboard.EndTime > 0 && expiryTime > leaderboard.EndTime {
+			expiryTime = leaderboard.EndTime
+		}
 	}
 
 	query := `UPDATE leaderboard_record SET max_num_score = (max_num_score + $1) WHERE leaderboard_id = $2 AND owner_id = $3 AND expiry_time = $4`
@@ -198,19 +204,8 @@ func TournamentList(ctx context.Context, logger *zap.Logger, db *sql.DB, categor
 SELECT 
 id, sort_order, reset_schedule, metadata, create_time, category, description, duration, end_time, max_size, max_num_score, title, size, start_time
 FROM leaderboard
-WHERE duration > 0 AND start_time >= $1 AND end_time <= $2 AND category >= $3 AND category <= $4`
-
+WHERE duration > 0 AND category >= $1 AND category <= $2 AND start_time >= $3`
 	params := make([]interface{}, 0, 6)
-	if startTime >= 0 {
-		params = append(params, time.Unix(int64(startTime), 0).UTC())
-	} else {
-		params = append(params, time.Unix(0, 0).UTC())
-	}
-	if endTime >= 0 {
-		params = append(params, time.Unix(int64(endTime), 0).UTC())
-	} else {
-		params = append(params, time.Unix(0, 0).UTC())
-	}
 	if categoryStart >= 0 && categoryStart <= 127 {
 		params = append(params, categoryStart)
 	} else {
@@ -222,7 +217,26 @@ WHERE duration > 0 AND start_time >= $1 AND end_time <= $2 AND category >= $3 AN
 		params = append(params, 127)
 	}
 
-	// To ensure that there are more records, so the cursor is returned/
+	if startTime >= 0 {
+		params = append(params, time.Unix(int64(startTime), 0).UTC())
+	} else {
+		params = append(params, time.Unix(0, 0).UTC())
+	}
+
+	if endTime == 0 {
+		query += " AND end_time = $4"
+		params = append(params, time.Unix(0, 0).UTC())
+	} else if int64(endTime) < now.Unix() {
+		// if end time is set explicitly in the past
+		query += " AND end_time <= $4"
+		params = append(params, time.Unix(int64(endTime), 0).UTC())
+	} else {
+		// if end time is in the future, return both tournaments that end in the future as well as the ones that never end.
+		query += " AND (end_time <= $4 OR end_time = '1970-01-01 00:00:00 UTC')"
+		params = append(params, time.Unix(int64(endTime), 0).UTC())
+	}
+
+	// To ensure that there are more records, so the cursor is returned
 	params = append(params, limit+1)
 
 	if cursor != nil {
@@ -449,7 +463,7 @@ func TournamentRecordWrite(ctx context.Context, logger *zap.Logger, db *sql.DB, 
 	return record, nil
 }
 
-func TournamentRecordsHaystack(ctx context.Context, logger *zap.Logger, db *sql.DB, leaderboardCache LeaderboardCache, rankCache LeaderboardRankCache, leaderboardId string, ownerId uuid.UUID, limit int) ([]*api.LeaderboardRecord, error) {
+func TournamentRecordsHaystack(ctx context.Context, logger *zap.Logger, db *sql.DB, leaderboardCache LeaderboardCache, rankCache LeaderboardRankCache, leaderboardId string, ownerId uuid.UUID, limit int, expiryOverride int64) ([]*api.LeaderboardRecord, error) {
 	leaderboard := leaderboardCache.Get(leaderboardId)
 	if leaderboard == nil {
 		return nil, ErrLeaderboardNotFound
@@ -457,9 +471,17 @@ func TournamentRecordsHaystack(ctx context.Context, logger *zap.Logger, db *sql.
 
 	sortOrder := leaderboard.SortOrder
 
-	_, _, expiry := calculateTournamentDeadlines(leaderboard.StartTime, leaderboard.EndTime, int64(leaderboard.Duration), leaderboard.ResetSchedule, time.Now().UTC())
-	expiryTime := time.Unix(expiry, 0).UTC()
+	expiry := expiryOverride
+	if expiry == 0 {
+		now := time.Now().UTC()
+		_, _, expiry = calculateTournamentDeadlines(leaderboard.StartTime, leaderboard.EndTime, int64(leaderboard.Duration), leaderboard.ResetSchedule, now)
+		if expiry != 0 && expiry <= now.Unix() {
+			// if the expiry time is in the past, we wont have any records to return
+			return make([]*api.LeaderboardRecord, 0), nil
+		}
+	}
 
+	expiryTime := time.Unix(expiry, 0).UTC()
 	return getLeaderboardRecordsHaystack(ctx, logger, db, rankCache, ownerId, limit, leaderboard.Id, sortOrder, expiryTime)
 }
 
@@ -486,6 +508,10 @@ func calculateTournamentDeadlines(startTime, endTime, duration int64, resetSched
 			expiryUnix = startActiveUnix + (schedule1Unix - schedule0Unix)
 		} else if startTime > startActiveUnix {
 			startActiveUnix = startTime
+		}
+
+		if endTime > 0 && expiryUnix > endTime {
+			expiryUnix = endTime
 		}
 
 		return startActiveUnix, endActiveUnix, expiryUnix